Fix determining the value from needle positions
Read the needle position values in precision order starting from the
most precise needle (on the left) and make +1 or -1 adjustment to each
value if the needle position does not match with the previous (more
precise) needle.  Use the adjusted value of the previous needle when
doing this checking if the value of the needle matches with previous
needle value.

This makes a big difference especially when the value is stepping over
from 9999 to 0000.