Fix caching of params related data
Both functions `get_dial_data` and `get_dials_template` cache their
result for speedups, but they're result is now depending on the contents
of the passed in params and a single global variable is not enough
anymore (since commit e64dd049).

Fix the caching by caching the results to a dictionary, which uses
`id(params)` as its key.